change to for loop, add instruments file, change data type
